requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
pandas
plotly
torch
torch-geometric
scikit-learn
joblib
numpy
pyyaml